Split coordinates.basis into basis matrix and local coordinate operation modules
