+++ mupt/geometry/coordinates/basis.py
@@ -1,11 +1,16 @@
-'''Calculations of local coordinate systems and linear bases'''
+'''Calculations to determine if linear bases enjoy certian properties, such as orthogonality'''
 
 __author__ = 'Timotej Bernat'
 __email__ = 'timotej.bernat@colorado.edu'
 
 import numpy as np
-from ..arraytypes import Shape, N, Dims, Numeric
+from ..arraytypes import Shape, N, Numeric
 
+
+def are_linearly_independent(*vectors: np.ndarray[Shape[N, ...], Numeric]) -> bool:
+    '''Check if the given set of vectors is linearly independent'''
+    # NOTE: numpy will raise Exception (as desired) when vectors passed have incompatible shapes
+    return np.linalg.matrix_rank(np.column_stack(vectors)) == len(vectors)
 
 def is_diagonal(matrix : np.ndarray[Shape[N, N], Numeric]) -> bool: # TODO: generalize to work for other diagonals
     '''Determine whether a matrix is digonal, i.e. has no nonzero elements off of the main diagonal'''
@@ -28,36 +33,3 @@ def is_orthogonal(matrix : np.ndarray[Shape[N, N], Numeric]) -> bool:
     return  np.allclose(matrix @ matrix.T, np.eye(n_rows, dtype=matrix.dtype)) \
         and np.allclose(matrix.T @ matrix, np.eye(n_cols, dtype=matrix.dtype)) # NOTE: can't optimize as the transpose of the above product for non-square matrices
 is_orthonormal = is_orthogonal
-
-def compute_local_coordinates(positions : np.ndarray[Shape[N, Dims], float]) -> tuple[
-        np.ndarray[Shape[Dims], float],
-        np.ndarray[Shape[Dims, Dims], float],
-        np.ndarray[Shape[Dims], float],
-    ]:
-    '''
-    Takes a coordinates vector of N D-dimensional points and determines
-    the center, axes, and relative lengths of the local principal coordinate systems
-    
-    Parameters
-    ----------
-    positions : Array[[N, D], float]
-        A vector of N points in D-dimensional
-    
-    Returns
-    -------
-    center : Array[[D,], float]
-        The D-dimensional coordinate point of the local coordinate origin
-    principal_axes : Array[[D, D], float]
-        A DxD matrix whose i-th column is the i-th basis vector in the local coordinate system
-        Basis provided is orthonormal (i.e. all columns have length 1 and are perpendicular to each other column)
-    axis_lengths : Array[[D,], float]
-        The relative length of each axis, if ordered by significance (i.e. amount of variation along that axis)
-    '''
-    center = positions.mean(axis=0)
-    
-    # determine principal axes from SVD
-    U, S, Vh = np.linalg.svd((positions - center), full_matrices=False) # NOTE: this places eigenvalues in descending order by default (no sorting needed)
-    principal_axes = eivecs = Vh.T          # transpose to place eigenvectors into column-order - NOTE: basis is guaranteed to be normal, since covariance matrix is real and symmetric
-    axis_lengths = eivals = (S * S) / (len(positions) - 1) # account for sample size normalization for covariance matrix
-    
-    return center, principal_axes, axis_lengths
